requirement.txt
Flask
Pillow
ultralytics
torch
